+++ source/renderer/app/stores/AddressesStore.js
@@ -42,7 +42,7 @@ export default class AddressesStore extends Store {
 
   @action _refreshAddresses = () => {
     if (this.stores.networkStatus.isConnected) {
-      const allWallets = this.stores.wallets.all;
+      const { allWallets } = this.stores.wallets;
       for (const wallet of allWallets) {
         const allRequest = this._getAddressesAllRequest(wallet.id);
         allRequest.invalidate({ immediately: false });