+++ source/renderer/app/stores/TransactionsStore.js
@@ -133,8 +133,8 @@ export default class TransactionsStore extends Store {
 
   @action _refreshTransactionData = () => {
     if (this.stores.networkStatus.isConnected) {
-      const allWallets = this.stores.wallets.all;
-      for (const wallet of allWallets) {
+      const { all: wallets } = this.stores.wallets;
+      for (const wallet of wallets) {
         const recentRequest = this._getTransactionsRecentRequest(wallet.id);
         recentRequest.execute({
           walletId: wallet.id,